+++ src/components/index.js
@@ -1,81 +1,40 @@
-import AggregateStatusCard from './AggregateStatusCard.vue';
-import ApplicationLauncher from './ApplicationLauncher.vue';
-import LauncherItem from './LauncherItem.vue';
-import Card from './Card.vue';
-import CardNotification from './CardNotification.vue';
-import Dropdown from './Dropdown.vue';
-import EmptyChart from './EmptyChart.vue';
-import Sort from './Sort.vue';
-import ColumnPicker from './ColumnPicker.vue';
-import Toolbar from './Toolbar.vue';
-import UtilizationBarChart from './UtilizationBarChart.vue';
-import FilterFields from './FilterFields.vue';
-import FilterResults from './FilterResults.vue';
-import Notifications from './Notifications.vue';
-import Notification from './Notification.vue';
-import NotificationBell from './NotificationBell.vue';
-import Drawer from './Drawer.vue';
-import DrawerGroup from './DrawerGroup.vue';
-import DrawerGroupAction from './DrawerGroupAction.vue';
-import DrawerNotification from './DrawerNotification.vue';
-import ExpandCollapse from './ExpandCollapse.vue';
-import Layout from './Layout.vue';
-import Table from './Table.vue';
-import PaginateControl from './PaginateControl.vue';
-import VerticalNavDivider from './VerticalNavDivider.vue';
-import VerticalSubmenu from './VerticalSubmenu.vue';
-import RadioButton from './RadioButton.vue';
-import Toggle from './Toggle.vue';
-import Combobox from './Combobox.vue';
-import Spinner from './Spinner.vue';
-import Select from './Select.vue';
-import Option from './Option.vue';
-import ListView from './ListView.vue';
-import ListItem from './ListItem.vue';
-import ListGroupItem from './ListGroupItem.vue';
-import ListItemAdditionalInfo from './ListItemAdditionalInfo.vue';
-import Modal from './Modal.vue';
-import Icon from './Icon';
-import MenuItem from './MenuItem';
-
-export {
-  AggregateStatusCard,
-  ApplicationLauncher,
-  LauncherItem,
-  Card,
-  CardNotification,
-  Dropdown,
-  EmptyChart,
-  Sort,
-  ColumnPicker,
-  Toolbar,
-  UtilizationBarChart,
-  FilterFields,
-  FilterResults,
-  Notifications,
-  Notification,
-  NotificationBell,
-  Drawer,
-  DrawerGroup,
-  DrawerGroupAction,
-  DrawerNotification,
-  ExpandCollapse,
-  Layout,
-  Table,
-  PaginateControl,
-  VerticalNavDivider,
-  VerticalSubmenu,
-  RadioButton,
-  Toggle,
-  Combobox,
-  Spinner,
-  Select,
-  Option,
-  ListView,
-  ListItem,
-  ListGroupItem,
-  ListItemAdditionalInfo,
-  Modal,
-  Icon,
-  MenuItem,
-};
+export { default as AggregateStatusCard } from './AggregateStatusCard.vue';
+export { default as ApplicationLauncher } from './ApplicationLauncher.vue';
+export { default as Button } from './Button.vue';
+export { default as LauncherItem } from './LauncherItem.vue';
+export { default as Card } from './Card.vue';
+export { default as CardNotification } from './CardNotification.vue';
+export { default as Dropdown } from './Dropdown.vue';
+export { default as EmptyChart } from './EmptyChart.vue';
+export { default as Sort } from './Sort.vue';
+export { default as ColumnPicker } from './ColumnPicker.vue';
+export { default as Toolbar } from './Toolbar.vue';
+export { default as UtilizationBarChart } from './UtilizationBarChart.vue';
+export { default as FilterFields } from './FilterFields.vue';
+export { default as FilterResults } from './FilterResults.vue';
+export { default as Notifications } from './Notifications.vue';
+export { default as Notification } from './Notification.vue';
+export { default as NotificationBell } from './NotificationBell.vue';
+export { default as Drawer } from './Drawer.vue';
+export { default as DrawerGroup } from './DrawerGroup.vue';
+export { default as DrawerGroupAction } from './DrawerGroupAction.vue';
+export { default as DrawerNotification } from './DrawerNotification.vue';
+export { default as ExpandCollapse } from './ExpandCollapse.vue';
+export { default as Layout } from './Layout.vue';
+export { default as Table } from './Table.vue';
+export { default as PaginateControl } from './PaginateControl.vue';
+export { default as VerticalNavDivider } from './VerticalNavDivider.vue';
+export { default as VerticalSubmenu } from './VerticalSubmenu.vue';
+export { default as RadioButton } from './RadioButton.vue';
+export { default as Toggle } from './Toggle.vue';
+export { default as Combobox } from './Combobox.vue';
+export { default as Spinner } from './Spinner.vue';
+export { default as Select } from './Select.vue';
+export { default as Option } from './Option.vue';
+export { default as ListView } from './ListView.vue';
+export { default as ListItem } from './ListItem.vue';
+export { default as ListGroupItem } from './ListGroupItem.vue';
+export { default as ListItemAdditionalInfo } from './ListItemAdditionalInfo.vue';
+export { default as Modal } from './Modal.vue';
+export { default as Icon } from './Icon';
+export { default as MenuItem } from './MenuItem';